requirements.txt
flask>=3.0
flask-socketio>=5.3
python-dotenv>=1.0
openai>=1.0
jinja2>=3.1
pyyaml>=6.0
markdown>=3.5
